requirements.txt
torch>=1.8.0
torchvision>=0.9.0
opencv-python
numpy
Pillow
tk
detectron2 @ git+https://github.com/facebookresearch/detectron2.git
